Extract default theme to separate file
